requirements.txt
django
ultralytics
opencv-python
celery
redis
djangorestframework
